+++ src/main/java/blackjack/domain/gamer/GameParticipant.java
@@ -1,4 +1,4 @@
-package blackjack.gamer;
+package blackjack.domain.gamer;
 
 import blackjack.domain.GameRule;
 import blackjack.domain.card.Card;
