+++ src/main/java/blackjack/domain/gamer/GameParticipants.java
@@ -1,4 +1,4 @@
-package blackjack.gamer;
+package blackjack.domain.gamer;
 
 import blackjack.domain.GameRule;
