+++ src/main/java/blackjack/domain/gamer/Nickname.java
@@ -1,4 +1,4 @@
-package blackjack.gamer;
+package blackjack.domain.gamer;
 
 import java.util.Objects;
