+++ src/test/java/blackjack/gamer/GameParticipantFixture.java
@@ -1,5 +1,10 @@
 package blackjack.gamer;
 
+import blackjack.domain.gamer.Dealer;
+import blackjack.domain.gamer.GameParticipant;
+import blackjack.domain.gamer.Nickname;
+import blackjack.domain.gamer.Player;
+
 import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Function;
