+++ src/test/java/blackjack/gamer/GameParticipantTest.java
@@ -1,6 +1,7 @@
 package blackjack.gamer;
 
 import blackjack.domain.card.CardFixture;
+import blackjack.domain.gamer.GameParticipant;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
