+++ src/test/java/blackjack/gamer/GameParticipantsTest.java
@@ -2,6 +2,10 @@ package blackjack.gamer;
 
 import blackjack.ConstantFixture;
 import blackjack.domain.GameRule;
+import blackjack.domain.gamer.Dealer;
+import blackjack.domain.gamer.GameParticipant;
+import blackjack.domain.gamer.GameParticipants;
+import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
