+++ src/test/java/blackjack/gamer/PlayerTest.java
@@ -1,5 +1,6 @@
 package blackjack.gamer;
 
+import blackjack.domain.gamer.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
